Decode torrent file information
